Refactor out route calculator
