+++ Tests/kps/business/RouteCalculatorTests.java
@@ -11,17 +11,19 @@ import org.junit.Test;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 
-public class RouteCalculator {
+public class RouteCalculatorTests {
     @Test public void ensureThatSystemCanCalculateARouteFromTestData() throws FileNotFoundException, XMLException {
         Simulation s = SimulationXML.readSimulationFromFile(new FileInputStream("test_data/Test.xml"));
-        CalculatedRoute calculatedRoutes = s.buildCalculatedRoute("Auckland", "Suva", Priority.INTERNATIONAL_AIR);
+        RouteCalculator rc = new RouteCalculator(s);
+        CalculatedRoute calculatedRoutes = rc.buildCalculatedRoute("Auckland", "Suva", Priority.INTERNATIONAL_AIR);
         Assert.assertNotNull(calculatedRoutes);
         Assert.assertTrue(calculatedRoutes.getRoutes().size() > 0);
     }
 
     @Test public void ensureThatSystemWillNotPickInferiorRoute() throws FileNotFoundException, XMLException {
         Simulation s = SimulationXML.readSimulationFromFile(new FileInputStream("test_data/Test.xml"));
-        CalculatedRoute calculatedRoute = s.buildCalculatedRoute("Auckland", "Suva", Priority.INTERNATIONAL_STANDARD);
+        RouteCalculator rc = new RouteCalculator(s);
+        CalculatedRoute calculatedRoute = rc.buildCalculatedRoute("Auckland", "Suva", Priority.INTERNATIONAL_STANDARD);
         Assert.assertTrue(calculatedRoute.getRoutes().size() == 3);
     }
